requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
Bootstrap-Flask==2.3.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
distro==1.9.0
docx==0.2.4
dominate==2.9.1
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
git-ignore==1.0.2
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
load-dotenv==0.1.0
lxml==6.0.2
MarkupSafe==3.0.2
openai==2.6.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
pydantic==2.12.3
pydantic_core==2.41.4
PyPDF2==3.0.1
python-dotenv==1.1.1
reportlab==4.4.4
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
visitor==0.1.3
Werkzeug==3.0.0
WTForms==3.0.1
